GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Virtual branch
   branch head: ca61cac47156ae3444f9851a0a12d88dc34a868d
   - themes/ianthedesigner-hugo/layouts/shopify/list.html
   - themes/ianthedesigner-hugo/layouts/dave-customer-success-workspace/list.html
   - themes/ianthedesigner-hugo/layouts/about/list.html
   - data/about.yml
   - content/about/_index.md

Your previous branch was: refs/heads/master

The sha for that commit was: 3a71584a87c8599297cfb8b3adbd111883780aa9

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch